refactor: separated function at OmokGameActivity
