+++ app/src/main/java/woowacourse/omok/view/activity/MainActivity.kt
@@ -1,11 +1,11 @@
-package woowacourse.omok
+package woowacourse.omok.view.activity
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
-import woowacourse.omok.view.OmokGameActivity
+import woowacourse.omok.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
